backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
qdrant-client
redis
aiofiles
neo4j
langchain
langchain-community
langchain-qdrant
langchain-ollama
langchain-huggingface
sentence-transformers
langchain-experimental
pypdf
tiktoken
python-dotenv
pydantic-settings
docling
duckduckgo-search
langgraph
bcrypt<4.0.0
geopy  # ✅ 좌표 변환용
tabulate # ✅ SQL 결과 출력용
pymysql # ✅ MySQL T2SQL 지원
cryptography # ✅ API 키 Fernet 암호화
# pinecone # (optional) Pinecone VDB 사용 시: pip install pinecone langchain-pinecone